+++ BugsnagPerformance/Assets/BugsnagPerformance/Scripts/Internal/BugsnagPerformanceAppLifecycleListener.cs
@@ -1,0 +1,28 @@
+using UnityEngine;
+namespace BugsnagUnityPerformance
+{
+    public class BugsnagPerformanceAppLifecycleListener : MonoBehaviour
+    {
+        void Awake()
+        {
+            DontDestroyOnLoad(gameObject);
+        }
+
+        void OnApplicationFocus(bool hasFocus)
+        {
+            if (!hasFocus)
+            {
+                BugsnagPerformance.AppBackgrounded();
+            }
+        }
+
+        void OnApplicationPause(bool paused)
+        {
+            if (paused)
+            {
+                BugsnagPerformance.AppBackgrounded();
+            }
+        }
+
+    }
+}