+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v1.3.4 (2024-04-29)
+
+### Bug Fixes
+
+- Fixed issue where spans could not be ended with a custom end time. [#106](https://github.com/bugsnag/bugsnag-unity-performance/pull/106)
+
+- Fixed issue where custom spans with the SpanOption IsFirstClass set to false still had it reported as true. [#107](https://github.com/bugsnag/bugsnag-unity-performance/pull/107)
+
+
 ## v1.3.3 (2024-03-07)
 
 ### Bug Fixes